+++ common/parse.h
@@ -46,5 +46,6 @@ int parseRawFormatAirmar(char *msg, RawMessage *m, bool showJson);
 int parseRawFormatChetco(char *msg, RawMessage *m, bool showJson);
 int parseRawFormatGarminCSV(char *msg, RawMessage *m, bool showJson, bool absolute);
 int parseRawFormatYDWG02(char *msg, RawMessage *m, bool showJson);
+int parseRawFormatActisenseN2KAscii(char *msg, RawMessage *m, bool showJson);
 
 #endif